Don't show itemtip if there was no content defined
With the old ternary operator assignment of content, it would always "item.children('td:last')" if item wasn't a li element. The problem with that is if no itemtip content was defined (td.ui-autocomplete-itemtip-content) then it would just dump the same autocomplete result content into the itemtip. I didn't think this looked very good so now it checks if td:last has the itemtip content class and if it doesn't, hides the itemtip and then returns early.