Catch errors when resetting Google reCaptcha
With the shutdown of reCaptcha API v1 the behaviour of the reset-function changed. If there is no p:captcha component on the page it will now throw an error and kill any subsequent javascript calls. This can be a problem if an ajax-request causes a forward or navigation to a new page where no captcha exists.

This change will catch the error and log it using PrimeFaces.info for a short information and PrimeFaces.debug if anyone is actually interested in the stack trace / error itself.