Fix primefaces#6575: selection could not be restored for lazy-mode + MVS
This fix adds two flags to restoreMultiViewState which are only relevant
for components that do not have a static data model but use lazy loaded
data. In that case the first flag indicates, that the data independent
state should be restored (e.g. filters in a datatable that is required
to load the data model). The second flag indicates, that the data is
available and the data dependent state could be restored.