Add integration test for issue #8883
Add integration test DatePicker014Test that uses an own date time class
and an own DateTimeConverter class.

The converter is bound from the data bean to the f:dateTimeConvert tag.
Normally one would instead configuring it globally in the
faces-config.xml (But I failed to do so in the integration test).