Fix #8883 Own DateTimeConverter should be considered (#8884)
* Fix #8883 Own DateTimeConverter should be considered

Converter.getType() has nothing to do with the type of the value that
the converter can handle. Possible values are for type are 'date',
'time' and 'both'.

Instead, I try to use the converter and if it is not able to convert a
value I put it on a blacklist, so that the next time the Primefaces
codes is used. This is important since the standard DateTimeConverter
does not handle java.time.* classes correct. On the other hand our own
converter should be used for our own date time data types.

* Fix #8883 Fix code style and indentation errors

* Add integration test for issue #8883

Add integration test DatePicker014Test that uses an own date time class
and an own DateTimeConverter class.

The converter is bound from the data bean to the f:dateTimeConvert tag.
Normally one would instead configuring it globally in the
faces-config.xml (But I failed to do so in the integration test).

* Fix #8883: Rename bad/good to invalid/valid and use app scoped map

Rename addBadConverter to addInvalidConverter
Rename isGoodConverter to isValidConverter
Instead of using a global Map to store the invalid converters use an app
scoped map.