10849: Charts tuple data interface
Defined a new Abstract Chart point and translation to json in the ChartUtils that has 3 sample implementations for Numeric x values, String x values and Date x values.  Updated the cartesian chart types to accept a list of type Object instead of Number to allow for use of the new Point implementations.  WIP implementation in the showcase Mixed Chart.