Tooltip autoHide="false": fix tooltip hidden when not crossing ui-tooltip-arrow
This fixes the check for allowHide by checking if the previous element's (the one that the mouse is leaving from) "aria-describedby" value is the same as the target element's (tooltip) or any of its parents' id (because if there is any child elements on the tooltip, the event may refer to it as the target instead of the tooltip itself). This includes the "ui-tooltip-arrow".