Fix PrimeFaces.setCaretToEnd
`PrimeFaces.setCaretToEnd` is broken. It takes a parameter `element`. Then it calls `element.trigger(...)`, which only works if `element` is a JQuery instance. Afterwards, it calls `element.setSelectionRange(...)`, which only works if `element` is an `HTMLInputElement` or `HTMLTextAreaElement`. No matter what you pass the method, one of these is going to fail.

The intention of the method seems to have been to take a raw HTMLInputElement or HTMLTextAreaElement. Use `element.focus()` instead, and remove `element.createTextRange` since that is an Internet Explorer-only method and we don't support IE anymore.